+++ SQLiteStudio3/guiSQLiteStudio/windows/tablewindow.cpp
@@ -565,8 +565,12 @@ void TableWindow::initDbAndTable()
     connect(ui->strictTableCheck, SIGNAL(clicked()), this, SLOT(strictChanged()));
 
     parseDdl();
+
     updateIndexes();
+    ui->indexList->sortByColumn(0, Qt::AscendingOrder);
+
     updateTriggers();
+    ui->triggerList->sortByColumn(0, Qt::AscendingOrder);
 
     // (Re)connect to DB signals
     connect(db, SIGNAL(dbObjectDeleted(QString,QString,DbObjectType)), this, SLOT(checkIfTableDeleted(QString,QString,DbObjectType)));
@@ -1544,6 +1548,7 @@ void TableWindow::updateIndexes()
     resolver.setIgnoreSystemObjects(false);
     QList<SqliteCreateIndexPtr> indexes = resolver.getParsedIndexesForTable(database, table);
 
+    ui->indexList->setSortingEnabled(false);
     ui->indexList->setColumnCount(4);
     ui->indexList->setRowCount(indexes.size());
     ui->indexList->setHorizontalHeaderLabels({
@@ -1582,6 +1587,9 @@ void TableWindow::updateIndexes()
 
     ui->indexList->resizeColumnsToContents();
     ui->indexList->horizontalHeader()->setSectionResizeMode(2, QHeaderView::Stretch);
+    ui->indexList->selectionModel()->clearSelection();
+    ui->indexList->selectionModel()->clearCurrentIndex();
+    ui->indexList->setSortingEnabled(true);
     updateIndexesState();
 }
 
@@ -1593,6 +1601,7 @@ void TableWindow::updateTriggers()
     SchemaResolver resolver(db);
     QList<SqliteCreateTriggerPtr> triggers = resolver.getParsedTriggersForTable(database, table);
 
+    ui->triggerList->setSortingEnabled(false);
     ui->triggerList->setColumnCount(4);
     ui->triggerList->setRowCount(triggers.size());
     ui->triggerList->horizontalHeader()->setMaximumSectionSize(200);
@@ -1630,6 +1639,9 @@ void TableWindow::updateTriggers()
 
     ui->triggerList->resizeColumnsToContents();
     ui->triggerList->horizontalHeader()->setSectionResizeMode(3, QHeaderView::Stretch);
+    ui->triggerList->selectionModel()->clearSelection();
+    ui->triggerList->selectionModel()->clearCurrentIndex();
+    ui->triggerList->setSortingEnabled(true);
     updateTriggersState();
 }